requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
sentence-transformers
chromadb
tavily-python
pypdf